Create NthNodeOfLoopedLL.java
Given a linked list with a loop. You need to find and return the Nth node from the point where the loop starts in the input LL traveling backward (as we move towards the head of the linked list).

Loop is created by connecting the last element of LL (i.e. tail) to some other node which is given in input.
The counting of nodes starts from 1.
Return null if no such element exists.