Create HappyNumber.java
Given a number n, find if it is a happy number or not. A number is called happy if a sequence of steps leads to 1 wherein each sequence of steps the number is replaced by the sum of the square of its digit, i.e., if we start with Happy Number and keep replacing it with digits square sum, we reach 1.