Create LaLiga.java
This is Super - Sunday for football fans. It is the final night. It is Barcelona vs Real Madrid. Moreover, the talking point of this year's La Liga is Sevilla becoming the last team to qualify for the playoffs with just 12 points.

The question is if 'N' teams are playing in the La Liga, what is the minimum number of points at which a team may qualify for a playoff?
Note: 1. You have to work around a solution in the premises of the format, where a winning team is awarded 2 points and the losing team is awarded 0 points. 4 teams qualify for the playoffs, after the league stage where each team plays all the other teams once.
2. We have to assume all the matches arrive at a result. (no draw)