Create BobAndHisString.java
King Bob is in a playful mood today. He started playing with string S. As he was playing, a weird question came to his mind. He wondered what is the maximum number of characters, between any two same characters in the string. He needs your help in solving this question. Can you help him solve this question?

Note: String S is composed of lowercase letters of the Latin alphabet. If there are no two same characters in the string, print -1.