Update Islands.java
An island is a small piece of land surrounded by water. A group of islands is said to be connected if we can reach from any given island to any other island in the same group. Given V islands (numbered from 1 to V) and E connections or edges between islands. Can you count the number of connected groups of islands?

Input Format :
The first line of input contains two integers, that denote the value of V and E.
Each of the following E lines contains two integers, that denote that there exists an edge between vertex a and b.