Create 3_cycle.java
Given a graph with N vertices (numbered from 0 to N-1) and M undirected edges, then count the distinct 3-cycles in the graph. A 3-cycle PQR is a cycle in which (P, Q), (Q, R), and (R, P) are connected by an edge.

Input Format :
The first line of input contains two space-separated integers, that denote the value of N and M.
Each of the following M lines contains two integers, that denote the vertices which have an undirected edge between them. Let us denote the two vertices with the symbol u and v.