Create LargestPiece.java
It's Gary's birthday today and he has ordered his favorite square cake consisting of '0's and '1's. But Gary wants the biggest piece of '1's and no '0's. A piece of cake is defined as a part that consists of only '1's, and all '1's share an edge with each other on the cake. Given the size of cake N and the cake, can you find the count of '1's in the biggest piece of '1's for Gary?

Input Format :
The first line of input contains an integer, that denotes the value of N. 
Each of the following N lines contains N space-separated integers.