Create PathToString.java
Given an NxM matrix containing Uppercase English Alphabets only. Your task is to tell if there is a path in the given matrix which makes the sentence “CODINGNINJA”.
There is a path from any cell to all its neighboring cells. For a particular cell, neighboring cells are those cells that share an edge or a corner with the cell.

Input Format :
The first line of input contains two space-separated integers N and M, where N is the number of rows and M is the number of columns in the matrix. 
Each of the following N lines contains M characters. Please note that characters are not space-separated.