Create ConnectingDots.java
Gary has a board of size NxM. Each cell on the board is a colored dot. There exist only 26 colors denoted by uppercase Latin characters (i.e. A, B,..., Z). Now Gary is getting bored and wants to play a game. The key of this game is to find a cycle that contains dots of the same color. Formally, we call a sequence of dots d1, d2, ..., dk a cycle if and only if it meets the following condition:
1. These k dots are different: if i ≠ j then di is different from dj.
2. k is at least 4.
3. All dots belong to the same color.
4. For all 1 ≤ i ≤ k - 1: di and di + 1 are adjacent. Also, dk and d1 should also be adjacent. Cells x and y are called adjacent if they share an edge.

Since Gary is color blind, he wants your help. Your task is to determine if there exists a cycle on the board.