Create KSmallestElements.java
You are given an integer k and an array of integers that contain numbers in random order. Write a program to find k smallest numbers from the given array. You need to save them in an array and return it.
Time complexity should be O(n * logk) and space complexity should not be more than O(k).

Note: The order of elements in the output is not important.

Input Format:
The first line of input contains an integer, that denotes the value of the size of the array. Let us denote it with the symbol N.
The following line contains N space-separated integers, that denote the value of the elements of the array.
The following line contains an integer, that denotes the value of k.