Create MaximumInvitations.java
Raj is throwing a birthday party. He wants to invite as many friends as possible. He wants a large wooden table to accommodate his friends. There is a constraint for this: space in the room. The number of friends a table can accommodate is equal to its perimeter (the sum of all side lengths). The furniture must be placed so that the edges are parallel to the edges of the flat.
You will be given the layout of the flat and you have to tell the maximum number of friends that he can accommodate in his room.

Input Format:
The first line of input contains two integers R and C (1 <= R, C <= 400), these are the dimensions of the flat.
Each of the following R rows contains exactly C characters. Characters can be '.', 'X'. '.' character means that cell is free and can be used for placing wooden table. 'X' means that the cell is blocked and hence, cannot be used.

Sample Input:
4 4
X.XX
X..X
..X.
..XX 

Sample Output:
9