Create GennadyAndHisExplosionString.java
Gennady likes to play with strings of characters, but this time he has taken it too far – he put an “explosion” in the string! An explosion is a series of characters that, if found in the vicinity of the fire, explodes and starts a chain reaction.
Gennady, being as negligent as he usually is, forgot that his string contained an explosion and placed it near candlelight. Thus the chain reaction began.

The chain reaction takes place in the following way:
• if a string contains explosions, they all explode and a new string is formed by concatenating the pieces without the exploding parts
• this concatenation could possibly create new explosions
• the chain reaction repeats while there are explosions in the string

Now Gennady wants to know whether anything will be left after this series of chain reactions. If nothing remains, output “FRULA” (without quotes). If, by any chance, something is left, output the final string remaining after all the reactions.

Please note: The explosion will not contain two equal characters.

Input 1:
mirkovC4nizCC44
C4 

Output 1:
mirkovniz 

Input 2:
12ab112ab2ab
12ab 

Output 2:
FRULA